Working on the get user by id (email)
It looks like the .route() method only takes /users/:id if not the actual users page -- ie. you can't give input of an email here. I will be changing this tomorrow so that the default id field is overwritten with the email so it makes for a simpler execution. Committing all changes for now, but this is not working code. (Not able to access a user by email).